Implement ExplodedStage component and refactor ExplodedBoard for improved exploded view functionality
- Introduced the `ExplodedStage` component to manage the rendering of exploded views, enhancing the separation of concerns in the exploded view logic.
- Refactored the `ExplodedBoard` component to utilize `ExplodedStage`, streamlining the process of displaying the anchored variant.
- Updated the exploded view logic to improve the handling of workspace changes and rendering performance.
- Enhanced styling and constants related to the exploded view, ensuring a cohesive visual representation.
- Added new utility functions and constants to support the exploded view's functionality and improve maintainability.
- Updated documentation and comments to clarify the new structure and behavior of the exploded view components.
